% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLCS.R
\name{createLCS}
\alias{createLCS}
\title{Create Latent Change Score Model}
\usage{
createLCS(num.obs, autoregression = 0.3, residualerrorvariance = 0.1,
  slopevariance = 0.1, interceptvariance = 0.1, has.slope = TRUE,
  has.icept = TRUE)
}
\description{
Create Latent Change Score Model
}

