% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRELCS.R
\name{createRELCS}
\alias{createRELCS}
\title{Create Random Effects Latent Change Score Model}
\usage{
createRELCS(num.obs, ar.variance = TRUE, first.obs.var.unique = FALSE,
  slope.ar.cov = FALSE, has.slope = TRUE, has.icept = TRUE)
}
\description{
Create Random Effects Latent Change Score Model
}

