% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateDataFromRELCS.R
\name{simulateDataFromRELCS}
\alias{simulateDataFromRELCS}
\title{Simulate Data from RELCS model}
\usage{
simulateDataFromRELCS(N, num.obs, autoregressionmean = 0.3,
  autoregressionvariance = 0.1, residualerrorvariance = 0.1,
  slopevariance = 0.1, interceptvariance = 0.1, has.slope = TRUE,
  has.icept = TRUE)
}
\description{
Simulate Data from RELCS model
}

