% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateDataFromRELCS.R
\name{simulateDataFromRELCS}
\alias{simulateDataFromRELCS}
\title{Simulate Data from RELCS model}
\usage{
simulateDataFromRELCS(N, num.obs, autoregressionmean = 0.3,
  autoregressionvariance = 0.1, residualerrorvariance = 0.1,
  slopevariance = 0, slopemu = 0, interceptvariance = 0.1,
  interceptmu = 4, has.slope = TRUE, has.icept = TRUE,
  use.openmx = FALSE)
}
\description{
Simulate Data from RELCS model
}

